+++ src/Routes/RoutesCommon.tsx
@@ -16,7 +16,7 @@ import { Farming } from 'pages/Defi/views/Farming'
 import { LiquidityPools } from 'pages/Defi/views/LiquidityPools'
 import { Overview } from 'pages/Defi/views/Overview'
 import { StakingVaults } from 'pages/Defi/views/StakingVaults'
-import { ProfileHeader } from 'pages/Profile/components/ProfileHeader'
+import { ProfileOverview } from 'pages/Profile/views/Overview'
 import { TransactionHistory } from 'pages/TransactionHistory/TransactionHistory'
 
 import { Route as NestedRoute } from './helpers'
@@ -141,7 +141,7 @@ export const routes: Array<NestedRoute> = [
   },
   {
     path: '/profile',
-    main: ProfileHeader,
+    main: ProfileOverview,
     hide: true,
     label: 'common.profile',
   },
